+++ package.json
@@ -23,6 +23,7 @@
   "dependencies": {
     "body-parser": "^1.19.0",
     "compression": "^1.7.4",
+    "helmet": "^4.2.0",
     "sanitize-html": "^2.3.0"
   }
 }
